Implement static portable Create methods for ECDsa
