+++ src/Common/src/Interop/Windows/Crypt32/Interop.HashIdAlg.cs
