+++ src/System.Security.Cryptography.Algorithms/tests/project.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "Microsoft.NETCore.Platforms": "1.0.1-rc3-24125-01",
     "System.IO": "4.1.0-rc3-24125-01",
+    "System.IO.FileSystem": "4.0.1-rc3-24125-01",
     "System.Linq.Expressions": "4.1.0-rc3-24125-01",
     "System.ObjectModel": "4.0.12-rc3-24125-01",
     "System.Runtime": "4.1.0-rc3-24125-01",
