+++ src/System.Security.Cryptography.Cng/src/System/Security/Cryptography/CngKey.Import.cs
@@ -26,22 +26,18 @@ namespace System.Security.Cryptography
             return Import(keyBlob, format, provider: CngProvider.MicrosoftSoftwareKeyStorageProvider);
         }
 
-#if !NETNATIVE
-        internal static CngKey Import(byte[] keyBlob, ECCurve curve, CngKeyBlobFormat format)
+        internal static CngKey Import(byte[] keyBlob, string curveName, CngKeyBlobFormat format)
         {
-            return Import(keyBlob, curve, format, provider: CngProvider.MicrosoftSoftwareKeyStorageProvider);
+            return Import(keyBlob, curveName, format, provider: CngProvider.MicrosoftSoftwareKeyStorageProvider);
         }
-#endif //!NETNATIVE
 
         public static CngKey Import(byte[] keyBlob, CngKeyBlobFormat format, CngProvider provider)
         {
-#if !NETNATIVE
             return Import(keyBlob, null, format, provider);
         }
 
-        internal static CngKey Import(byte[] keyBlob, ECCurve? curve, CngKeyBlobFormat format, CngProvider provider)
+        internal static CngKey Import(byte[] keyBlob, string curveName, CngKeyBlobFormat format, CngProvider provider)
         {
-#endif //!NETNATIVE
             if (keyBlob == null)
                 throw new ArgumentNullException(nameof(keyBlob));
             if (format == null)
@@ -50,12 +46,10 @@ namespace System.Security.Cryptography
                 throw new ArgumentNullException(nameof(provider));
 
             SafeNCryptProviderHandle providerHandle = provider.OpenStorageProvider();
-            SafeNCryptKeyHandle keyHandle;
+            SafeNCryptKeyHandle keyHandle = null;
             ErrorCode errorCode;
             
-#if !NETNATIVE
-            if (curve == null)
-#endif //!NETNATIVE
+            if (curveName == null)
             {
                 errorCode = Interop.NCrypt.NCryptImportKey(providerHandle, IntPtr.Zero, format.Format, IntPtr.Zero, out keyHandle, keyBlob, keyBlob.Length, 0);
                 if (errorCode != ErrorCode.ERROR_SUCCESS)
@@ -63,53 +57,12 @@ namespace System.Security.Cryptography
                     throw errorCode.ToCryptographicException();
                 }
             }
-#if !NETNATIVE
             else
             {
-                // Call with Oid.FriendlyName because .Value will result in an invalid parameter error
-                Debug.Assert(curve.Value.IsNamed);
-                string curveName = curve.Value.Oid.FriendlyName; 
-                using (SafeUnicodeStringHandle safeCurveName = new SafeUnicodeStringHandle(curveName))
-                {
-                    var desc = new Interop.BCrypt.BCryptBufferDesc();
-                    var buff = new Interop.BCrypt.BCryptBuffer();
-
-                    IntPtr descPtr = IntPtr.Zero;
-                    IntPtr buffPtr = IntPtr.Zero;
-                    try
-                    {
-                        descPtr = Marshal.AllocHGlobal(Marshal.SizeOf(desc));
-                        buffPtr = Marshal.AllocHGlobal(Marshal.SizeOf(buff));
-                        buff.cbBuffer = (curveName.Length + 1) * 2; // Add 1 for null terminator
-                        buff.BufferType = Interop.BCrypt.NCryptBufferDescriptors.NCRYPTBUFFER_ECC_CURVE_NAME;
-                        buff.pvBuffer = safeCurveName.DangerousGetHandle();
-                        Marshal.StructureToPtr(buff, buffPtr, false);
-
-                        desc.cBuffers = 1;
-                        desc.pBuffers = buffPtr;
-                        desc.ulVersion = Interop.BCrypt.BCRYPTBUFFER_VERSION;
-                        Marshal.StructureToPtr(desc, descPtr, false);
-
-                        errorCode = Interop.NCrypt.NCryptImportKey(providerHandle, IntPtr.Zero, format.Format, descPtr, out keyHandle, keyBlob, keyBlob.Length, 0);
-                    }
-                    finally
-                    {
-                        Marshal.FreeHGlobal(descPtr);
-                        Marshal.FreeHGlobal(buffPtr);
-                    }
-                }
-
-                if (errorCode != ErrorCode.ERROR_SUCCESS)
-                {
-                    Exception e = errorCode.ToCryptographicException();
-                    if (errorCode == ErrorCode.NTE_INVALID_PARAMETER)
-                    {
-                        throw new PlatformNotSupportedException(string.Format(SR.Cryptography_CurveNotSupported, curveName), e);
-                    }
-                    throw e;
-                }
-            }
+#if !NETNATIVE
+                keyHandle = ECCng.ImportKeyBlob(format.Format, keyBlob, curveName, providerHandle);
 #endif //!NETNATIVE
+            }
 
             CngKey key = new CngKey(providerHandle, keyHandle);
